# openapi.json
{
  "components": {
    "schemas": {
      "BulkStatusChange": {
        "additionalProperties": false,
        "properties": {
          "ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "maxItems": 500,
            "minItems": 1,
            "title": "Ids",
            "type": "array"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          }
        },
        "required": [
          "ids",
          "status"
        ],
        "title": "BulkStatusChange",
        "type": "object"
      },
      "BulkStatusResult": {
        "properties": {
          "not_found": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "title": "Not Found",
            "type": "array"
          },
          "unchanged": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "title": "Unchanged",
            "type": "array"
          },
          "updated": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "title": "Updated",
            "type": "array"
          }
        },
        "required": [
          "updated",
          "unchanged",
          "not_found"
        ],
        "title": "BulkStatusResult",
        "type": "object"
      },
...
      "CompanyMatches": {
        "properties": {
          "companies": {
            "items": {
              "$ref": "#/components/schemas/CompanyMatch"
            },
            "title": "Companies",
            "type": "array"
          }
        },
        "required": [
          "companies"
        ],
        "title": "CompanyMatches",
        "type": "object"
      },
...
      "DuplicateJob": {
        "properties": {
          "detail": {
            "default": "Already tracked",
            "title": "Detail",
            "type": "string"
          },
          "existing_id": {
            "format": "uuid",
            "title": "Existing Id",
            "type": "string"
          }
        },
        "required": [
          "existing_id"
        ],
        "title": "DuplicateJob",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "HistoryEntry": {
        "properties": {
          "changed_at": {
            "format": "date-time",
            "title": "Changed At",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          }
        },
        "required": [
          "status",
          "changed_at"
        ],
        "title": "HistoryEntry",
        "type": "object"
      },
...
      "JobCreate": {
        "additionalProperties": false,
        "properties": {
          "applied_at": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Applied At"
          },
          "company": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Company",
            "type": "string"
          },
          "contact_email": {
            "anyOf": [
              {
                "maxLength": 320,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contact Email"
          },
          "contact_name": {
            "anyOf": [
              {
                "maxLength": 320,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contact Name"
          },
          "location": {
            "anyOf": [
              {
                "maxLength": 320,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "notes": {
            "anyOf": [
              {
                "maxLength": 10000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "role": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Role",
            "type": "string"
          },
          "salary": {
            "anyOf": [
              {
                "maxLength": 320,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Salary"
          },
          "source": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JobSource"
              },
              {
                "type": "null"
              }
            ]
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus",
            "default": "saved"
          },
          "url": {
            "anyOf": [
              {
                "maxLength": 2048,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          }
        },
        "required": [
          "company",
          "role"
        ],
        "title": "JobCreate",
        "type": "object"
      },
      "JobOut": {
        "properties": {
          "applied_at": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Applied At"
          },
          "attention": {
            "anyOf": [
              {
                "enum": [
                  "needs_follow_up",
                  "still_to_apply",
                  "no_response"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Attention"
          },
          "company": {
            "title": "Company",
            "type": "string"
          },
          "contact_email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contact Email"
          },
          "contact_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contact Name"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "days_since_last_change": {
            "title": "Days Since Last Change",
            "type": "integer"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "last_status_change_at": {
            "format": "date-time",
            "title": "Last Status Change At",
            "type": "string"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "role": {
            "title": "Role",
            "type": "string"
          },
          "salary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Salary"
          },
          "source": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JobSource"
              },
              {
                "type": "null"
              }
            ]
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          }
        },
        "required": [
          "id",
          "company",
          "role",
          "url",
          "location",
          "source",
          "salary",
          "contact_name",
          "contact_email",
          "status",
          "applied_at",
          "notes",
          "created_at",
          "updated_at",
          "last_status_change_at",
          "days_since_last_change",
          "attention"
        ],
        "title": "JobOut",
        "type": "object"
      },
      "JobPage": {
        "properties": {
          "counts": {
            "additionalProperties": {
              "type": "integer"
            },
            "propertyNames": {
              "$ref": "#/components/schemas/JobStatus"
            },
            "title": "Counts",
            "type": "object"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/JobOut"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "title": "Page Size",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "page_size",
          "counts"
        ],
        "title": "JobPage",
        "type": "object"
      },
      "JobSource": {
        "enum": [
          "linkedin",
          "company_site",
          "referral",
          "recruiter",
          "indeed",
          "other"
        ],
        "title": "JobSource",
        "type": "string"
      },
      "JobStatus": {
        "enum": [
          "saved",
          "applied",
          "interviewing",
          "offer",
          "accepted",
          "rejected",
          "withdrawn",
          "no_response"
        ],
        "title": "JobStatus",
        "type": "string"
      },
...
      "StatusChange": {
        "additionalProperties": false,
        "properties": {
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          }
        },
        "required": [
          "status"
        ],
        "title": "StatusChange",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Job Tracker",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/jobs": {
      "get": {
        "description": "Jobs matching the filter and search, one page at a time.\n\nThe filter is a status, or a group: `active` (saved, applied, interviewing, offer),\n`closed` (rejected, withdrawn, no response) or `all`.\n\n`q` searches company and role (case-insensitive, at least 2 characters). `counts`\ngives every status's total across all your jobs, ignoring the filter and search.",
        "operationId": "list_jobs_api_v1_jobs_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "active",
                    "closed",
                    "all"
                  ],
                  "type": "string"
                },
                {
                  "$ref": "#/components/schemas/JobStatus"
                }
              ],
              "default": "active",
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "default": "created_at",
              "enum": [
                "company",
                "role",
                "status",
                "last_status_change",
                "applied_at",
                "created_at"
              ],
              "title": "Sort",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "order",
            "required": false,
            "schema": {
              "default": "desc",
              "enum": [
                "asc",
                "desc"
              ],
              "title": "Order",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 25,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Jobs",
        "tags": [
          "jobs"
        ]
      },
      "post": {
        "description": "Create a job. 409 with the existing job's id if you already track this URL.",
        "operationId": "create_job_api_v1_jobs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DuplicateJob"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Job",
        "tags": [
          "jobs"
        ]
      }
    },
    "/api/v1/jobs/bulk-status": {
      "post": {
        "description": "Set many jobs' status at once (e.g. \"Mark all as no response\"), all or nothing.\n\nOnly your own jobs are changed: any other ID is reported in `not_found` and never\ntouched. Jobs already in the status are reported in `unchanged`.",
        "operationId": "bulk_change_status_api_v1_jobs_bulk_status_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkStatusChange"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulkStatusResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Bulk Change Status",
        "tags": [
          "jobs"
        ]
      }
    },
    "/api/v1/jobs/company-matches": {
      "get": {
        "description": "Your jobs at companies whose name matches, by the start of a word.\n\nCase, punctuation and legal suffixes (Ltd, Limited, Inc\u2026) are ignored when matching,\nso \"acme\" finds \"ACME\" and \"Acme Ltd\", and \"ac\" also finds \"Accenture\" (but not\n\"Pacific\"). Each name is returned as entered, with its own counts: nothing is merged.",
        "operationId": "company_matches_api_v1_jobs_company_matches_get",
        "parameters": [
          {
            "in": "query",
            "name": "company",
            "required": true,
            "schema": {
              "title": "Company",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanyMatches"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Company Matches",
        "tags": [
          "jobs"
        ]
      }
    },
    "/api/v1/jobs/{job_id}": {
      "delete": {
        "description": "Delete a job permanently, with its status history and interviews.",
        "operationId": "delete_job_api_v1_jobs__job_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Job",
        "tags": [
          "jobs"
        ]
      },
      "get": {
        "operationId": "get_job_api_v1_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Job",
        "tags": [
          "jobs"
        ]
      },
      "patch": {
        "description": "Change only the fields sent; null clears an optional field. Not the status (JT-26).",
        "operationId": "update_job_api_v1_jobs__job_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DuplicateJob"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Job",
        "tags": [
          "jobs"
        ]
      }
    },
    "/api/v1/jobs/{job_id}/history": {
      "get": {
        "description": "The job's status changes, newest first.",
        "operationId": "status_history_api_v1_jobs__job_id__history_get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/HistoryEntry"
                  },
                  "title": "Response Status History Api V1 Jobs  Job Id  History Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Status History",
        "tags": [
          "jobs"
        ]
      }
    },
...
    "/api/v1/jobs/{job_id}/status": {
      "post": {
        "description": "Change a job's status, recording it in the history.\n\nAny status can move to any other. Setting the status it already has changes nothing.\nBecoming `applied` fills in today's date (in your time zone) if there's none;\ngoing back to `saved` clears it.",
        "operationId": "change_status_api_v1_jobs__job_id__status_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StatusChange"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Change Status",
        "tags": [
          "jobs"
        ]
      }
    },
...
  }
}